backend/requirements.txt
fastapi
uvicorn
openai
spacy
sentence-transformers
scikit-learn
python-dotenv
